+++ lib/Latte/Nodes/ScriptNode.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace JanHerman\Barista\Latte\Nodes;
+
+class ScriptNode extends SfcNode
+{
+    protected const Languages = ['js', 'ts'];
+}