+++ lib/Latte/Nodes/StyleNode.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace JanHerman\Barista\Latte\Nodes;
+
+class StyleNode extends SfcNode
+{
+    protected const Languages = ['css', 'scss'];
+}